disable logic tests, remove not needed core tests, fix basic core tests
