+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "test": "npm run clarinet:test",
     "test:token": "npm run clarinet:prepare && clarinet test ./tests/citycoin-token.test.ts",
-    "test:core": "npm run clarinet:prepare && clarinet test ./tests/citycoin-core.test.ts",
+    "test:core": "npm run clarinet:prepare && clarinet test ./tests/citycoin-core-v1.test.ts",
     "clarinet:check": "npm run clarinet:prepare && clarinet check",
     "clarinet:test": "npm run clarinet:prepare && clarinet test",
     "clarinet:codecov": "npm run clarinet:prepare && clarinet test --coverage",
