+++ src/core-client.ts
@@ -39,7 +39,7 @@ export class CoreClient extends Client {
   unsafeSetCityWallet(newCityWallet: Account): Tx {
     return Tx.contractCall(
       this.contractName,
-      "unsafe-set-city-wallet",
+      "test-unsafe-set-city-wallet",
       [types.principal(newCityWallet.address)],
       this.deployer.address
     );
@@ -56,61 +56,5 @@ export class CoreClient extends Client {
 
   getCityWallet(): ReadOnlyFn {
     return this.callReadOnlyFn("get-city-wallet");
-  }
-
-  proposeContract(name: string, contractAddress: string, sender: Account): Tx {
-    return Tx.contractCall(
-      this.contractName,
-      "propose-contract",
-      [types.ascii(name), types.principal(contractAddress)],
-      sender.address
-    );
-  }
-
-  getContract(contractAddress: string): ReadOnlyFn {
-    return this.callReadOnlyFn("get-contract", [
-      types.principal(contractAddress),
-    ]);
-  }
-
-  getProposal(id: number): ReadOnlyFn {
-    return this.callReadOnlyFn("get-proposal", [types.uint(id)]);
-  }
-
-  vote(proposalId: number | undefined, sender: Account): Tx {
-    return Tx.contractCall(
-      this.contractName,
-      "vote",
-      [
-        typeof proposalId === "undefined"
-          ? types.none()
-          : types.some(types.uint(proposalId)),
-      ],
-      sender.address
-    );
-  }
-
-  closeProposal(id: number, sender: Account) {
-    return Tx.contractCall(
-      this.contractName,
-      "close-proposal",
-      [types.uint(id)],
-      sender.address
-    );
-  }
-
-  getActiveContract(name: string): ReadOnlyFn {
-    return this.callReadOnlyFn("get-active-contract", [types.ascii(name)]);
-  }
-
-  createProposalTuple(data: ProposalTuple): object {
-    return {
-      address: data.contractAddress,
-      startBH: types.uint(data.startBH),
-      endBH: types.uint(data.endBH),
-      voters: types.uint(data.voters),
-      votes: types.uint(data.votes),
-      isOpen: types.bool(data.isOpen),
-    };
   }
 }
